Add API version flag and disable logging file flag
Summary:
## Add a config flag for specifying API version

Until we figure out how to hit the latest API version automatically. This gives us "quicker fixes" for clients who don't update the code.

Adding API version as a parameter to `graphAPI`, and most of the changes are around function signatures and tests.

## Add a config flag for disabling writing log files to disk

This is a work-around for a bug `winston` has on Windows currently where file creation throws even if run as administrator.

## Updated jest version to latest

There were bunch of security errors before updating

## Intentionally *not* updating README

Since these are advertiser-specific fixes, don't want to complicate the doc further.

Reviewed By: yuetaoxu

Differential Revision: D10510849

fbshipit-source-id: 9beafe8d