Surface API error message on failed calls
Summary:
Currently, the implementation of `graphAPI` only bubbles up the API response if the status code is `200`. However, we found that this makes it difficult for users to debug. For instance, [this](https://github.com/facebookincubator/offline-conversion-file-uploader/issues/11#issuecomment-440134322) github issue.

This diff also extracts responses from failed API calls and wrap it into the `Error`.

Reviewed By: yuetaoxu

Differential Revision: D13144128

fbshipit-source-id: f328b65f