added api version to readme file for offline conversions
Summary:
Adding apiVersion to Readme.md since for some reason it fails for me if I don't include it.

it seems apiVersion is now required otherwise the script can't work it out itself.

Currently, it tries to call this endpoint: https://graph.facebook.com/api_version?access_token=xxx
However, it doesn't work. This is because this endpoint was deprecated on 30 April in D15082623.

Reviewed By: supasate

Differential Revision: D15516045

fbshipit-source-id: b0c4b7d7